Class type Protocol_client_context.full

The class type full allows to create contexts that are explicitly used by low-level shell functions, while containing various information (I/O services, RPCs...). Then, depending on the usage, the type may be coerced into one of its following ascendants to serve for explicit operations on blocks, chain or daemon for instance.

inherit Tezos_client_base.Client_context.full

inherit (Tezos_shell_services.Shell_services.chain * Tezos_shell_services.Shell_services.block) Tezos_protocol_008_PtEdo2Zk.Protocol.Environment.RPC_context.simple

inherit [Tezos_shell_services.Shell_services.chain, Tezos_shell_services.Shell_services.block] Tezos_protocol_008_PtEdo2Zk.Protocol.Environment.proto_rpc_context