Module Tezos_client_008_PtEdo2Zk__Managed_contract

val get_contract_manager : Tezos_client_008_PtEdo2Zk.Protocol_client_context.#full -> Tezos_raw_protocol_008_PtEdo2Zk.Alpha_context.Contract.t -> Tezos_raw_protocol_008_PtEdo2Zk.Alpha_context.public_key_hash Tezos_base__TzPervasives.tzresult Lwt.t

Retrieve the manager key in a contract storage. The storage has to be of type `pair key_hash 'a`.

val build_delegate_operation : Tezos_client_008_PtEdo2Zk.Protocol_client_context.#full -> chain:Tezos_shell_services.Chain_services.chain -> block:Tezos_shell_services.Block_services.block -> ?⁠fee:Tezos_raw_protocol_008_PtEdo2Zk.Alpha_context.Tez.t -> Tezos_raw_protocol_008_PtEdo2Zk.Alpha_context.Contract.t -> Tezos_raw_protocol_008_PtEdo2Zk.Alpha_context.public_key_hash option -> Tezos_raw_protocol_008_PtEdo2Zk.Alpha_context.Kind.transaction Tezos_client_008_PtEdo2Zk.Injection.annotated_manager_operation Tezos_base__TzPervasives.tzresult Lwt.t

Builds a delegation operation ready for injection

val set_delegate : Tezos_client_008_PtEdo2Zk.Protocol_client_context.#full -> chain:Tezos_shell_services.Chain_services.chain -> block:Tezos_shell_services.Block_services.block -> ?⁠confirmations:int -> ?⁠dry_run:bool -> ?⁠verbose_signing:bool -> ?⁠simulation:bool -> ?⁠branch:int -> fee_parameter:Tezos_client_008_PtEdo2Zk.Injection.fee_parameter -> ?⁠fee:Tezos_raw_protocol_008_PtEdo2Zk.Alpha_context.Tez.t -> source:Tezos_raw_protocol_008_PtEdo2Zk.Alpha_context.public_key_hash -> src_pk:Tezos_raw_protocol_008_PtEdo2Zk.Alpha_context.public_key -> src_sk:Tezos_client_base.Client_keys.sk_uri -> Tezos_raw_protocol_008_PtEdo2Zk.Alpha_context.Contract.t -> Tezos_raw_protocol_008_PtEdo2Zk.Alpha_context.public_key_hash option -> Tezos_raw_protocol_008_PtEdo2Zk.Alpha_context.Kind.transaction Tezos_raw_protocol_008_PtEdo2Zk.Alpha_context.Kind.manager Tezos_client_008_PtEdo2Zk.Injection.result Tezos_base__TzPervasives.tzresult Lwt.t

Set the delegate of a manageable contract. For a contract with a `do`entrypoint, it builds the lambda that set the provided delegate. `~source` has to be the registered manager of the contract.

val build_transaction_operation : Tezos_client_008_PtEdo2Zk.Protocol_client_context.#full -> chain:Tezos_shell_services.Chain_services.chain -> block:Tezos_shell_services.Block_services.block -> contract:Tezos_raw_protocol_008_PtEdo2Zk.Alpha_context.Contract.t -> destination:Tezos_raw_protocol_008_PtEdo2Zk.Alpha_context.Contract.t -> ?⁠entrypoint:string -> ?⁠arg:string -> amount:Tezos_raw_protocol_008_PtEdo2Zk.Alpha_context.Tez.t -> ?⁠fee:Tezos_raw_protocol_008_PtEdo2Zk.Alpha_context.Tez.t -> ?⁠gas_limit:Tezos_raw_protocol_008_PtEdo2Zk.Alpha_context.Gas.Arith.integral -> ?⁠storage_limit:Tezos_raw_protocol_008_PtEdo2Zk.Alpha_context.counter -> unit -> Tezos_raw_protocol_008_PtEdo2Zk.Alpha_context.Kind.transaction Tezos_client_008_PtEdo2Zk.Injection.annotated_manager_operation Tezos_base__TzPervasives.tzresult Lwt.t

Builds a transaction operation ready for injection

val transfer : Tezos_client_008_PtEdo2Zk.Protocol_client_context.#full -> chain:Tezos_shell_services.Chain_services.chain -> block:Tezos_shell_services.Block_services.block -> ?⁠confirmations:int -> ?⁠dry_run:bool -> ?⁠verbose_signing:bool -> ?⁠simulation:bool -> ?⁠branch:int -> source:Tezos_raw_protocol_008_PtEdo2Zk.Alpha_context.public_key_hash -> src_pk:Tezos_raw_protocol_008_PtEdo2Zk.Alpha_context.public_key -> src_sk:Tezos_client_base.Client_keys.sk_uri -> contract:Tezos_raw_protocol_008_PtEdo2Zk.Alpha_context.Contract.t -> destination:Tezos_raw_protocol_008_PtEdo2Zk.Alpha_context.Contract.t -> ?⁠entrypoint:string -> ?⁠arg:string -> amount:Tezos_raw_protocol_008_PtEdo2Zk.Alpha_context.Tez.t -> ?⁠fee:Tezos_raw_protocol_008_PtEdo2Zk.Alpha_context.Tez.t -> ?⁠gas_limit:Tezos_raw_protocol_008_PtEdo2Zk.Alpha_context.Gas.Arith.integral -> ?⁠storage_limit:Tezos_raw_protocol_008_PtEdo2Zk.Alpha_context.counter -> ?⁠counter:Tezos_raw_protocol_008_PtEdo2Zk.Alpha_context.counter -> fee_parameter:Tezos_client_008_PtEdo2Zk.Injection.fee_parameter -> unit -> (Tezos_raw_protocol_008_PtEdo2Zk.Alpha_context.Kind.transaction Tezos_raw_protocol_008_PtEdo2Zk.Alpha_context.Kind.manager Tezos_client_008_PtEdo2Zk.Injection.result * Tezos_raw_protocol_008_PtEdo2Zk.Alpha_context.Contract.t list) Tezos_base__TzPervasives.tzresult Lwt.t

Perform a transfer on behalf of a managed contract . For a contract with a `do`entrypoint, it builds the lambda that does the requested operation. `~source` has to be the registered manager of the contract.

val build_lambda_for_set_delegate : delegate:Tezos_raw_protocol_008_PtEdo2Zk.Alpha_context.public_key_hash option -> string

val build_lambda_for_transfer_to_implicit : destination:Tezos_raw_protocol_008_PtEdo2Zk.Alpha_context.public_key_hash -> amount:Tezos_raw_protocol_008_PtEdo2Zk.Alpha_context.Tez.t -> string

val build_lambda_for_transfer_to_originated : destination:Tezos_raw_protocol_008_PtEdo2Zk.Alpha_context.Contract.t -> entrypoint:string -> amount:Tezos_raw_protocol_008_PtEdo2Zk.Alpha_context.Tez.t -> parameter_type:Tezos_raw_protocol_008_PtEdo2Zk.Alpha_context.Script.expr -> parameter:Tezos_raw_protocol_008_PtEdo2Zk.Alpha_context.Script.expr -> string